# querymiddleware: Pool snappy writer in shard activity series