#### What this PR does

In https://github.com/grafana/mimir/pull/7290#pullrequestreview-1862270232 @pracucci suggested adding snappy writer to the existing benchmarks of the middleware. This PR splits `BenchmarkActiveSeriesMiddlewareMergeResponses` into two categories: with and without snappy encoding.

After comparing the with/without results, and since it seems that, shard activity series is a part of some warm path in GEM, I decided it's worth it to pool snappy writer. Here are the results before and after the optimization:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/frontend/querymiddleware
                                                                         │ /Users/v/tmp/bench-1.txt │      /Users/v/tmp/bench-2a.txt      │
                                                                         │          sec/op          │    sec/op     vs base               │
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-2-4                 33.25µ ± ∞ ¹   13.34µ ± ∞ ¹  -59.87% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-4-4                 33.75µ ± ∞ ¹   13.86µ ± ∞ ¹  -58.95% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-8-4                 35.84µ ± ∞ ¹   15.37µ ± ∞ ¹  -57.13% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-16-4                41.02µ ± ∞ ¹   20.87µ ± ∞ ¹  -49.13% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-32-4                50.65µ ± ∞ ¹   29.13µ ± ∞ ¹  -42.48% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-64-4                71.37µ ± ∞ ¹   45.73µ ± ∞ ¹  -35.93% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-128-4              106.21µ ± ∞ ¹   73.56µ ± ∞ ¹  -30.74% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-256-4               175.9µ ± ∞ ¹   138.6µ ± ∞ ¹  -21.19% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-512-4               308.8µ ± ∞ ¹   247.4µ ± ∞ ¹  -19.88% (p=0.008 n=5)
geomean                                                                                68.77µ         38.80µ        -43.58%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                         │ /Users/v/tmp/bench-1.txt │      /Users/v/tmp/bench-2a.txt       │
                                                                         │           B/op           │     B/op       vs base               │
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-2-4             2058.624Ki ± ∞ ¹   1.940Ki ± ∞ ¹  -99.91% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-4-4             2058.843Ki ± ∞ ¹   2.132Ki ± ∞ ¹  -99.90% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-8-4             2059.312Ki ± ∞ ¹   2.627Ki ± ∞ ¹  -99.87% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-16-4            2060.255Ki ± ∞ ¹   3.674Ki ± ∞ ¹  -99.82% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-32-4            2062.128Ki ± ∞ ¹   5.537Ki ± ∞ ¹  -99.73% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-64-4            2065.896Ki ± ∞ ¹   9.436Ki ± ∞ ¹  -99.54% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-128-4            2073.57Ki ± ∞ ¹   17.32Ki ± ∞ ¹  -99.16% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-256-4             2171.1Ki ± ∞ ¹   115.9Ki ± ∞ ¹  -94.66% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-512-4             2201.7Ki ± ∞ ¹   148.5Ki ± ∞ ¹  -93.26% (p=0.008 n=5)
geomean                                                                               2.041Mi         9.485Ki        -99.55%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                         │ /Users/v/tmp/bench-1.txt │      /Users/v/tmp/bench-2a.txt      │
                                                                         │        allocs/op         │  allocs/op    vs base               │
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-2-4                  47.00 ± ∞ ¹    40.00 ± ∞ ¹  -14.89% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-4-4                  65.00 ± ∞ ¹    58.00 ± ∞ ¹  -10.77% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-8-4                 101.00 ± ∞ ¹    94.00 ± ∞ ¹   -6.93% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-16-4                 173.0 ± ∞ ¹    166.0 ± ∞ ¹   -4.05% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-32-4                 317.0 ± ∞ ¹    310.0 ± ∞ ¹   -2.21% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-64-4                 605.0 ± ∞ ¹    598.0 ± ∞ ¹   -1.16% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-128-4               1.181k ± ∞ ¹   1.174k ± ∞ ¹   -0.59% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-256-4               2.346k ± ∞ ¹   2.339k ± ∞ ¹   -0.30% (p=0.008 n=5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-512-4               4.650k ± ∞ ¹   4.643k ± ∞ ¹   -0.15% (p=0.008 n=5)
geomean                                                                                 370.6          353.2         -4.69%
¹ need >= 6 samples for confidence interval at level 0.95
```

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.